Renamed a few packages to internal in the sonos binding. See openhab/static-code-analysis#100
