+++ extensions/binding/org.eclipse.smarthome.binding.sonos/src/main/java/org/eclipse/smarthome/binding/sonos/internal/SonosHandlerFactory.java
@@ -7,15 +7,14 @@
  */
 package org.eclipse.smarthome.binding.sonos.internal;
 
-import static org.eclipse.smarthome.binding.sonos.config.ZonePlayerConfiguration.UDN;
+import static org.eclipse.smarthome.binding.sonos.internal.config.ZonePlayerConfiguration.UDN;
 
 import java.util.Dictionary;
 import java.util.Hashtable;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
-import org.eclipse.smarthome.binding.sonos.SonosBindingConstants;
-import org.eclipse.smarthome.binding.sonos.handler.ZonePlayerHandler;
+import org.eclipse.smarthome.binding.sonos.internal.handler.ZonePlayerHandler;
 import org.eclipse.smarthome.config.core.Configuration;
 import org.eclipse.smarthome.config.discovery.DiscoveryServiceRegistry;
 import org.eclipse.smarthome.core.audio.AudioHTTPServer;
