+++ extensions/binding/org.eclipse.smarthome.binding.sonos/src/main/java/org/eclipse/smarthome/binding/sonos/internal/config/ZonePlayerConfiguration.java
@@ -5,7 +5,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.eclipse.smarthome.binding.sonos.config;
+package org.eclipse.smarthome.binding.sonos.internal.config;
 
 /**
  *
